requirements.txt
# Core dependencies
anthropic>=0.18.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.9.0
aiohttp>=3.9.0
aiofiles>=23.2.0
jinja2>=3.1.0
pyyaml>=6.0.0
